--- src/components/OpcionRespuesta/OpcionRespuesta.css
@@ -1,67 +1,0 @@
-.respuesta {
-  padding: 30px 30px;
-  font-weight: bold;
-  font-family: 'Poppins', sans-serif;
-  border: none;
-  border-radius: 15px;
-  cursor: pointer;
-  transition: all 0.1s ease-in-out;
-  position: relative;
-}
-  
-  .respuesta:hover {
-    filter: brightness(1.1);
-  }
-  
-  .verde {
-    background-color: #B5DF71;
-    color: #484800;
-    box-shadow: 0 6px 0 #8EBA47;
-  }
-
-  .verde:active {
-    box-shadow: 0 2px 0 #8EBA47;
-    transform: translateY(4px);
-  }
-  
-  .rojo {
-    background-color: #F29B5E;
-    color: #622C07;
-    box-shadow: 0 6px 0 #FF7544;
-  }
-  .rojo:active {
-    box-shadow: 0 2px 0 #FF7544;
-    transform: translateY(4px);
-  }
-  
-  .amarillo {
-    background-color: #FFD578;
-    color: #544013;
-    box-shadow: 0 6px 0 #E9B647;
-    
-  }
-  .amarillo:active {
-    box-shadow: 0 2px 0 #E9B647;
-    transform: translateY(4px);
-  }
-  
-  .azul {
-    background-color: #9EE6F4;
-    color: #124F5B;
-    box-shadow: 0 6px 0 #73CDDF;
-  }
-  .azul:active {
-    box-shadow: 0 2px 0 #73CDDF;
-    transform: translateY(4px);
-  }
-  
-  .correcta {
-    background-color: #22c55e;
-    color: white;
-  }
-  
-  .incorrecta {
-    background-color: #ef4444;
-    color: white;
-  }
-
